fix(aap-85119): extract pollExecutionStatus to shared API utility
